bot/requirements.txt
# LuckyRed Bot Dependencies
python-telegram-bot==20.7
sqlalchemy==2.0.25
asyncpg==0.29.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
httpx==0.26.0
aiohttp==3.9.1
